Rename feature to `Comments` from `Notes`
https://github.com/metasfresh/metasfresh/issues/6302
