+++ de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5557830_sys_gh6504_RenameToCommentsFromNotes.sql
@@ -1,0 +1,20 @@
+-- 2020-04-23T11:47:52.252Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Message SET MsgText='Kommentare', Value='webui.window.comments.caption',Updated=TO_TIMESTAMP('2020-04-23 14:47:52','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Message_ID=544978
+;
+
+-- 2020-04-23T11:47:56.687Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Message_Trl SET MsgText='Kommentare ',Updated=TO_TIMESTAMP('2020-04-23 14:47:56','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='de_CH' AND AD_Message_ID=544978
+;
+
+-- 2020-04-23T11:48:01.965Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Message_Trl SET MsgText='Comments',Updated=TO_TIMESTAMP('2020-04-23 14:48:01','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='nl_NL' AND AD_Message_ID=544978
+;
+
+-- 2020-04-23T11:48:04.192Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Message_Trl SET MsgText='Comments',Updated=TO_TIMESTAMP('2020-04-23 14:48:04','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='en_US' AND AD_Message_ID=544978
+;
+
